Fix tests, fix handling of security.enabled flag, clean up and organize new classes
