+++ api/jpo-conflictvisualizer-api/src/main/java/us/dot/its/jpo/ode/api/keycloak/support/KeycloakJwtAuthenticationConverter.java
@@ -1,4 +1,4 @@
-package us.dot.its.jpo.ode.api.keycloak;
+package us.dot.its.jpo.ode.api.keycloak.support;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.convert.converter.Converter;
