+++ api/jpo-conflictvisualizer-api/src/test/java/us/dot/its/jpo/ode/api/BsmTest.java
@@ -35,7 +35,7 @@ public class BsmTest {
   @Test
   public void testBsmJson() {
 
-    //MockKeyCloakAuth.setSecurityContextHolder("cm_user", Set.of("USER"));
+    MockKeyCloakAuth.setSecurityContextHolder("cm_user", Set.of("USER"));
 
     List<OdeBsmData> list = new ArrayList<>();
