+++ api/jpo-conflictvisualizer-api/src/test/java/us/dot/its/jpo/ode/api/MockKeyCloakAuth.java
@@ -1,27 +1,33 @@
 package us.dot.its.jpo.ode.api;
 
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
 import java.security.Principal;
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Set;
 
-//import org.keycloak.adapters.OidcKeycloakAccount;
-//import org.keycloak.adapters.springsecurity.token.KeycloakAuthenticationToken;
+
+import org.mockito.Mockito;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
+
+import static org.mockito.Mockito.*;
 
 public class MockKeyCloakAuth {
     public static void setSecurityContextHolder(String user, Set<String> roles){
         Principal principal = mock(Principal.class);
         when(principal.getName()).thenReturn(user);
 
-//        OidcKeycloakAccount account = mock(OidcKeycloakAccount.class);
-//        when(account.getRoles()).thenReturn(roles);
-//        when(account.getPrincipal()).thenReturn(principal);
-//
-//        KeycloakAuthenticationToken authentication = mock(KeycloakAuthenticationToken.class);
-//        when(authentication.getAccount()).thenReturn(account);
-//
-//        SecurityContextHolder.getContext().setAuthentication(authentication);
+        Authentication authentication = mock(Authentication.class);
+        when(authentication.isAuthenticated()).thenReturn(true);
+        Collection<GrantedAuthority> authorities = new ArrayList<>();
+        for (String role : roles) {
+            authorities.add(new SimpleGrantedAuthority("ROLE_" + role));
+        }
+        // type safe "when" doesn't work here
+        doReturn(authorities).when(authentication).getAuthorities();
+
+        SecurityContextHolder.getContext().setAuthentication(authentication);
     }
 }
